pyproject.toml
[project]
name = "cofy-api"
authors = [
    { name="EnergyID", email = "info@energieid.be" },
    { name="Jorg" }
]
version = "0.1.0"
description = "An open-source modular framework for ingesting, standardising, storing, and computing energy-related data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "entsoe-py>=0.7.10",
    "fastapi[standard]>=0.128.0",
    "narwhals>=2.15.0",
    "pandas>=3.0.0",
    "polars>=1.37.1",
    "saq[redis,web]>=0.26.1",
]
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"
[tool.setuptools.packages.find]
where = ["src"]
[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
    "ty>=0.0.12",
]
[tool.poe.tasks]
lint = "ruff check --fix"
format = "ruff format"
check = "ty check"
[tool.poe.tasks.demo]
cmd = "fastapi dev demo/main.py"
envfile = ".env.local"
[tool.poe.tasks.db-seed]
cmd = "python -m demo.db.seed"
envfile = ".env.local"
[tool.poe.tasks.db-reset]
cmd = "python -m demo.db.reset"
envfile = ".env.local"
[tool.poe.tasks.db-migrate]
cmd = "python -m demo.db.migrate"
envfile = ".env.local"
[tool.poe.tasks.db-generate]
cmd = "python -m demo.db.generate"
envfile = ".env.local"
[tool.poe.tasks.worker]
cmd = "saq demo.worker.settings"
envfile = ".env.local"
[tool.poe.tasks.test]
cmd = "pytest --cov=cofy --cov-report=term-missing --import-mode=importlib"
[tool.coverage.report]
skip_covered = true
[tool.coverage.run]
source = ["cofy"]
omit = ["cofy/modules/*/migrations/*", "cofy/db/alembic/*"]
[tool.ruff]
line-length = 120
[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
]
ignore = [
    "E501", #Line too long
]
